inference/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
onnxruntime-gpu
